feat: fix tests issues, fix default osm editor value on user profile page
