+++ src/features/osm_edit_link/constants.ts
@@ -2,4 +2,3 @@ import { i18n } from '~core/localization';
 
 export const EDIT_IN_OSM_CONTROL_ID = 'EditInOsm';
 export const EDIT_IN_OSM_CONTROL_NAME = i18n.t('toolbar.edit_in_osm');
-export const DEFAULT_OSM_EDITOR = 'osm';
